Add a sample analyzer project with dependencies to sanity-check
In order to test that the basics of the dotnet-ilrepack tool are working in all supported OSes, we pick a clean and common scenario that doesn't require any particular workload, and which builds fast, as a simple end-to-end sanity check for the global tool itself.

Note that the extensive tests already run on the code from existing tests is a prerequisite already for the test tool job.